# [Trace Exporters] Remove enumeration helper and use new net7 enumeration APIs