… src/OpenTelemetry.Api/Internal/EnumerationHelper.cs
… src/OpenTelemetry.Api/Internal/IActivityEnumerator.cs
… src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
… src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OpenTelemetry.Exporter.OpenTelemetryProtocol.csproj
… src/OpenTelemetry.Exporter.Zipkin/Implementation/ZipkinActivityConversionExtensions.cs
… src/OpenTelemetry.Exporter.Zipkin/OpenTelemetry.Exporter.Zipkin.csproj
… test/OpenTelemetry.Exporter.OpenTelemetryProtocol.Tests/OtlpTraceExporterTests.cs
… test/OpenTelemetry.Exporter.Zipkin.Tests/Implementation/ZipkinActivityConversionExtensionsTest.cs
… test/OpenTelemetry.Instrumentation.W3cTraceContext.Tests/OpenTelemetry.Instrumentation.W3cTraceContext.Tests.csproj
… test/OpenTelemetry.Tests/Trace/ActivityExtensionsTest.cs
… test/OpenTelemetry.Tests/Trace/LinkTest.cs