🥳

Also updated OtlpExporter to send "dropped counts" when attributes limits are being used. CHANGELOG note added for that.