@CodeBlanch 2022-08-18
@alanwest Let me run the benchmarks to make sure no allocations snuck in and I'll post the results.