@CodeBlanch 2022-08-18
Jaeger benchmarks:

(What this benchmark tests is that Jaeger reserves its memory and then reuses it for subsequent calls. Allocations should stay flat no matter the number of export calls.)

**Before**
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev | Allocated |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|----------:|
| JaegerExporter_Batching |               1 |         10000 |    30.81 ms |  0.066 ms |  0.062 ms |     16 KB |
| JaegerExporter_Batching |              10 |         10000 |   301.86 ms |  1.151 ms |  1.021 ms |     18 KB |
| JaegerExporter_Batching |             100 |         10000 | 3,073.62 ms | 46.019 ms | 40.795 ms |     18 KB |

**After**
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev | Allocated |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|----------:|
| JaegerExporter_Batching |               1 |         10000 |    29.88 ms |  0.142 ms |  0.126 ms |     16 KB |
| JaegerExporter_Batching |              10 |         10000 |   297.06 ms |  1.300 ms |  1.153 ms |     18 KB |
| JaegerExporter_Batching |             100 |         10000 | 3,021.04 ms | 19.152 ms | 14.952 ms |     18 KB |