@CodeBlanch 2022-08-18
Zipkin benchmarks:

(I am offended by how greedy Zipkin is with its allocations! 😱 This can be improved, which I will take as a TODO. The important thing is, no change as a result of this PR.)

**Before**
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev |      Median | Completed Work Items | Lock Contentions |      Gen 0 |     Gen 1 | Allocated |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|------------:|---------------------:|-----------------:|-----------:|----------:|----------:|
| ZipkinExporter_Batching |               1 |         10000 |    42.51 ms |  0.747 ms |  0.699 ms |    42.61 ms |               4.0000 |                - |   142.8571 |         - |      2 MB |
| ZipkinExporter_Batching |              10 |         10000 |   406.07 ms |  8.039 ms | 17.646 ms |   412.61 ms |              14.0000 |                - |  1000.0000 |         - |     17 MB |
| ZipkinExporter_Batching |             100 |         10000 | 4,132.82 ms | 10.160 ms |  9.503 ms | 4,128.53 ms |             104.0000 |                - | 14000.0000 | 1000.0000 |    169 MB |

**After**
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev | Completed Work Items | Lock Contentions |      Gen 0 |     Gen 1 | Allocated |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|---------------------:|-----------------:|-----------:|----------:|----------:|
| ZipkinExporter_Batching |               1 |         10000 |    37.42 ms |  0.134 ms |  0.118 ms |               4.0714 |                - |   142.8571 |         - |      2 MB |
| ZipkinExporter_Batching |              10 |         10000 |   368.16 ms |  2.066 ms |  1.832 ms |              13.0000 |                - |  1000.0000 |         - |     17 MB |
| ZipkinExporter_Batching |             100 |         10000 | 3,653.48 ms | 17.635 ms | 14.726 ms |             103.0000 |                - | 14000.0000 | 1000.0000 |    169 MB |