@CodeBlanch 2022-08-18
OTLP benchmarks:

(Some nice gains here* but the memory usage is still extremely high. By the way, it is not fair to compare the time of this benchmark against Jaeger or Zipkin above. OtlpExporter is not faster. The OtlpExporter benchmark mocks the whole call into Protobuf which owns the serialization phase. Jaeger and Zipkin benchmarks actually measure the serialization too. Safe to say OtlpExporter is slower and it likely also uses more memory than what the benchmark sees!)

**Before**
|                Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev |      Median |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|---------------------- |---------------- |-------------- |------------:|----------:|----------:|------------:|------------:|------------:|-----------:|----------:|
| OtlpExporter_Batching |               1 |         10000 |    36.73 ms |  0.699 ms |  1.902 ms |    35.98 ms |   2076.9231 |   1153.8462 |   384.6154 |     22 MB |
| OtlpExporter_Batching |              10 |         10000 |   366.33 ms |  7.284 ms | 15.042 ms |   364.83 ms |  21000.0000 |  12000.0000 |  4000.0000 |    217 MB |
| OtlpExporter_Batching |             100 |         10000 | 4,017.00 ms | 71.843 ms | 59.992 ms | 4,014.98 ms | 224000.0000 | 133000.0000 | 44000.0000 |  2,167 MB |

**After**
|                Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev |       Gen 0 |      Gen 1 |      Gen 2 | Allocated |
|---------------------- |---------------- |-------------- |------------:|----------:|----------:|------------:|-----------:|-----------:|----------:|
| OtlpExporter_Batching |               1 |         10000 |    21.05 ms |  0.258 ms |  0.242 ms |   1531.2500 |   781.2500 |   343.7500 |     18 MB |
| OtlpExporter_Batching |              10 |         10000 |   221.46 ms |  3.872 ms |  7.176 ms |  16000.0000 |  8000.0000 |  2000.0000 |    182 MB |
| OtlpExporter_Batching |             100 |         10000 | 2,086.82 ms | 26.596 ms | 24.878 ms | 153000.0000 | 77000.0000 | 37000.0000 |  1,824 MB |

\* The perf gains are due to the removal of `PooledList<T>` usage from OtlpExporter. `PooledList<T>` is not necessarily bad but how Protobuf was interacting with it was causing a lot of waste. This...

```csharp
    public sealed class RepeatedField<T>
    {
        public void AddRange(IEnumerable<T> values)
        {
            ProtoPreconditions.CheckNotNull(values, nameof(values));

            // Optimization 1: If the collection we're adding is already a RepeatedField<T>,
            // we know the values are valid.
            var otherRepeatedField = values as RepeatedField<T>;
            if (otherRepeatedField != null)
            {
                EnsureSize(count + otherRepeatedField.count);
                Array.Copy(otherRepeatedField.array, 0, array, count, otherRepeatedField.count);
                count += otherRepeatedField.count;
                return;
            }

            // Optimization 2: The collection is an ICollection, so we can expand
            // just once and ask the collection to copy itself into the array.
            var collection = values as ICollection;
            if (collection != null)
            {
                var extraCount = collection.Count;
                // For reference types and nullable value types, we need to check that there are no nulls
                // present. (This isn't a thread-safe approach, but we don't advertise this is thread-safe.)
                // We expect the JITter to optimize this test to true/false, so it's effectively conditional
                // specialization.
                if (default(T) == null)
                {
                    // TODO: Measure whether iterating once to check and then letting the collection copy
                    // itself is faster or slower than iterating and adding as we go. For large
                    // collections this will not be great in terms of cache usage... but the optimized
                    // copy may be significantly faster than doing it one at a time.
                    foreach (var item in collection)
                    {
                        if (item == null)
                        {
                            throw new ArgumentException("Sequence contained null element", nameof(values));
                        }
                    }
                }
                EnsureSize(count + extraCount);
                collection.CopyTo(array, count);
                count += extraCount;
                return;
            }

            // We *could* check for ICollection<T> as well, but very very few collections implement
            // ICollection<T> but not ICollection. (HashSet<T> does, for one...)

            // Fall back to a slower path of adding items one at a time.
            foreach (T item in values)
            {
                Add(item);
            }
        }
    }
```

Issues...

* `IEnumerable<T>` boxes the `struct PooledList<T>` and its enumerator. Leads to allocations.
* In the `ICollection` path Protobuf enumerates the source to look for `null` entries. That is a lot of CPU waste because we have already validated everything added.
* Just universally speaking, adding things to one array and copying them to another array is going to be slower then adding them to the final array directly 😄